[server] Moved ThirdEyePrincipal to server module
