+++ thirdeye-server/src/main/java/ai/startree/thirdeye/auth/ThirdEyeAuthenticatorDisabled.java
@@ -13,7 +13,6 @@
  */
 package ai.startree.thirdeye.auth;
 
-import ai.startree.thirdeye.spi.ThirdEyePrincipal;
 import io.dropwizard.auth.AuthenticationException;
 import io.dropwizard.auth.Authenticator;
 import java.util.Optional;
