+++ thirdeye-server/src/main/java/ai/startree/thirdeye/auth/oauth/ThirdEyeOAuthAuthenticator.java
@@ -15,7 +15,7 @@ package ai.startree.thirdeye.auth.oauth;
 
 import static ai.startree.thirdeye.spi.util.SpiUtils.optional;
 
-import ai.startree.thirdeye.spi.ThirdEyePrincipal;
+import ai.startree.thirdeye.auth.ThirdEyePrincipal;
 import com.google.common.cache.LoadingCache;
 import com.google.inject.Inject;
 import io.dropwizard.auth.AuthenticationException;
