+++ thirdeye-server/src/main/java/ai/startree/thirdeye/resources/AuthResource.java
@@ -16,7 +16,7 @@ package ai.startree.thirdeye.resources;
 import static ai.startree.thirdeye.spi.Constants.NO_AUTH_USER;
 import static ai.startree.thirdeye.util.ResourceUtils.respondOk;
 
-import ai.startree.thirdeye.spi.ThirdEyePrincipal;
+import ai.startree.thirdeye.auth.ThirdEyePrincipal;
 import ai.startree.thirdeye.spi.api.AuthApi;
 import ai.startree.thirdeye.spi.api.UserApi;
 import com.codahale.metrics.annotation.Timed;
